update: general update
    new file:   .config/dunst/dunstrc
        added dunst config not major changes from defaults
    modified:   .config/fish/conf.d/aliases.fish
        added: xdg-open now alias of handlr (https://github.com/chmln/handlr)
    modified:   .config/fish/conf.d/updt.fish
        update -> nvim_update shortcut (disablet it because of restart
        question)
        update -> dmacs_update added doom sync for fixes upgrade issues
    modified:   .config/hypr/apps.conf
        add -> added dbus-update-activation enviorment ...
               for export some wayland env vars
    modified:   .config/hypr/binds.conf
        update -> change binding mod+l to  exec rofi powermenu mode
    modified:   .config/hypr/variables.conf
        add -> wayland env vars
    new file:   .config/zathura/zathurarc
        add -> some basic shortcuts and auto clipboard selection
    new file:   .doom.d/config.org
        add -> config explanation and config tangle
    modified:   .doom.d/init.el
        add -> zig
    modified:   .firefox_scipts_configs/user.js
        update -> generally removed unnecessary and updated customization
    new file:   .local/bin/addvm.sh
        add -> add jake@linux vm creation script and updated it to my preferences
    new file:   .local/bin/rofi_mode_powermenu
        add -> patched version of https://github.com/jluttine/rofi-power-menu
    modified:   nix_packages.txt
        add -> updated it to current packages
    deleted:    scripts/audio.sh
        rm -> created better version (.loca/bin/rofi_audio)
    deleted:    scripts/batter_level.sh
        rm -> unnecessary

+++ .config/fish/conf.d/aliases.fish
@@ -11,7 +11,11 @@ if command -v fd >/dev/null
 end
 
 # mv, rm, cp
-alias rm="rm -i -v"
+if command -v trash > /dev/null
+	alias rm="trash"
+else 
+	alias rm="rm -i -v"
+end
 alias mv="mv -v"
 alias cp="cp -v"
 
@@ -37,3 +41,7 @@ alias vim="nvim"
 # prints all info about package fuzzy yay search installed "Q" and non-installed "S" database
 alias fzy_yaySs="yay -Ss | paste -d '' - - | fzf --preview 'echo {}' --preview-window down,10% | lolcat"
 alias fzy_yayQs="yay -Qs | paste -d '' - - | fzf --preview 'echo {}' --preview-window down,10% | lolcat"
+
+if command -v handlr >/dev/null
+   alias xdg-open="handlr"
+end
